Moved a lot of the logic in TabFriends to TabNav so It can also be accessed in TabDM
